MC-1460: add File.trimBox so that suppliers can easily see the flat dimentions of folded items
