+++ src/main/java/com/mixam/mxjdf4/sdk/Dimensions.java
@@ -1,0 +1,38 @@
+package com.mixam.mxjdf4.sdk;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Dimensions {
+
+    private double longEdge;
+    private double shortEdge;
+    private LengthUnitType units;
+
+    public double getLongEdge() {
+        return longEdge;
+    }
+
+    public void setLongEdge(double longEdge) {
+        this.longEdge = longEdge;
+    }
+
+    public double getShortEdge() {
+        return shortEdge;
+    }
+
+    public void setShortEdge(double shortEdge) {
+        this.shortEdge = shortEdge;
+    }
+
+    public LengthUnitType getUnits() {
+        return units;
+    }
+
+    public void setUnits(LengthUnitType units) {
+        this.units = units;
+    }
+
+}
