+++ src/main/java/com/mixam/mxjdf4/sdk/Format.java
@@ -5,36 +5,9 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
+public class Format extends Dimensions {
 
-public class Format {
-    private double longEdge;
     private OrientationType orientation;
-    private double shortEdge;
-    private LengthUnitType units;
-
-    public double getLongEdge() {
-        return longEdge;
-    }
-
-    public void setLongEdge(double longEdge) {
-        this.longEdge = longEdge;
-    }
-
-    public double getShortEdge() {
-        return shortEdge;
-    }
-
-    public void setShortEdge(double shortEdge) {
-        this.shortEdge = shortEdge;
-    }
-
-    public LengthUnitType getUnits() {
-        return units;
-    }
-
-    public void setUnits(LengthUnitType units) {
-        this.units = units;
-    }
 
     public OrientationType getOrientation() {
         return orientation;
@@ -43,6 +16,5 @@ public class Format {
     public void setOrientation(OrientationType orientation) {
         this.orientation = orientation;
     }
-
 
 }
